requirements.txt
aiogram>=3.15.0
pydantic>=2.4.1,<2.10
instructor>=1.3.0
openai>=1.14.0
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
